+++ RealState/Controllers/Public/PublicTicketController.cs
@@ -18,7 +18,7 @@ public class PublicTicketController(IUnitOfWork unitOfWork, JwtTokenService toke
 	public async Task<long?> GetCurrentUserId()
 	{
 		var userId = _tokenService.GetUserIdFromClaims(User) ?? "0";
-		var user = await _unitOfWork.UserRepository.GetUser(int.Parse(userId));
+		var user = await _unitOfWork.UserRepository.GetUser(long.Parse(userId));
 		return user?.id;
 	}
