+++ RealState/Controllers/Public/PublicUserPropertyController.cs
@@ -19,7 +19,7 @@ public class PublicUserPropertyController(IUnitOfWork unitOfWork, JwtTokenServic
 	public async Task<long?> GetCurrentUserId()
 	{
 		var userId = _tokenService.GetUserIdFromClaims(User) ?? "0";
-		var user = await _unitOfWork.UserRepository.GetUser(int.Parse(userId));
+		var user = await _unitOfWork.UserRepository.GetUser(long.Parse(userId));
 		return user?.id;
 	}
 	[HttpPost]
@@ -28,7 +28,7 @@ public class PublicUserPropertyController(IUnitOfWork unitOfWork, JwtTokenServic
 	public async Task<User?> GetCurrentUser()
 	{
 		var userId = _tokenService.GetUserIdFromClaims(User) ?? "0";
-		var user = await _unitOfWork.UserRepository.GetUser(int.Parse(userId));
+		var user = await _unitOfWork.UserRepository.GetUser(long.Parse(userId));
 		return user;
 	}
 	[HttpGet]
